Ingredients displays its' table with data from mysql
So this essentially replaces home.html, I ran it off of wamp and it was
able to work. I just used the setup scripts to configure the mysql
server and I was able to run the php We should refactor the dynamic
table into its' own function. Should work in user functionality as well.